feat(e2e-suite): add locally working native e2e reporter
DRY githunReporter methods
minor file cleanup
